upgraded this example to SDK 28 along with refactoring away from the old exponent
