+++ .babelrc
@@ -1,8 +1,12 @@
 {
-  "presets": ["babel-preset-exponent"],
+  "presets": [
+    "babel-preset-expo"
+  ],
   "env": {
     "development": {
-      "plugins": ["transform-react-jsx-source"]
+      "plugins": [
+        "transform-react-jsx-source"
+      ]
     }
   }
 }
