+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules/**/*
 .exponent/*
+.expo/*
 npm-debug.*
+.DS_Store
