--- exp.json
@@ -1,21 +1,0 @@
-{
-  "name": "video-background",
-  "description": "People like to make videos as backgrounds for authentication screens and other things, this shows how.",
-  "slug": "video-background",
-  "sdkVersion": "11.0.0",
-  "version": "1.0.0",
-  "orientation": "portrait",
-  "primaryColor": "#cccccc",
-  "iconUrl": "https://s3.amazonaws.com/exp-brand-assets/ExponentEmptyManifest_192.png",
-  "notification": {
-    "iconUrl": "https://s3.amazonaws.com/exp-us-standard/placeholder-push-icon-blue-circle.png",
-    "color": "#000000"
-  },
-  "loading": {
-    "iconUrl": "https://s3.amazonaws.com/exp-brand-assets/ExponentEmptyManifest_192.png",
-    "hideExponentText": false
-  },
-  "packagerOpts": {
-    "assetExts": ["ttf", "mp4"]
-  }
-}
