+++ package.json
@@ -4,11 +4,10 @@
   "description": "Hello Exponent!",
   "author": "exponent.team@gmail.com",
   "private": true,
-  "main": "main.js",
+  "main": "node_modules/expo/AppEntry.js",
   "dependencies": {
-    "exponent": "~11.0.2",
-    "@exponent/vector-icons": "~2.0.3",
-    "react": "~15.3.2",
-    "react-native": "git+https://github.com/exponentjs/react-native#sdk-11.0.3"
+    "expo": "^28.0.0",
+    "react": "16.3.1",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-28.0.0.tar.gz"
   }
 }
